requirements/requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
comtypes==1.1.9
cryptography==3.4.7
cx-Oracle==8.1.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
httmock==1.4.0
http-request-randomizer==1.3.2
httplib2==0.19.1
idna==2.10
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.20.2
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2021.1
pywin32==300
pywinauto==0.6.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
uritemplate==3.0.1
urllib3==1.26.4
xlrd==2.0.1
